==> 2.synonymize.ipynb <==
from nltk.tokenize import word_tokenize
from nltk.stem import WordNetLemmatizer 
from nltk.corpus import stopwords
import json
import nltk
from nltk import pos_tag
from nltk.corpus import wordnet
import re
import pickle
import pandas as pd
nltk.download('stopwords')
nltk.download('punkt')
from icecream import ic

with open('data/topics_lemmatized.txt', 'rb') as f:
    queries_lem = pickle.load(f)
f.close()

df = pd.DataFrame(columns = ['query', 'objs', 'syn'])

df['query'] = queries_lem
print(df)
df.to_csv('data/q_for_syn.tsv', sep = '\t')

## Detect words for synonyms manually, then read the file and find synonyms



df = pd.read_csv('data/q_for_syn.tsv', sep = '\t')
df = df.fillna(0)


#query ='difference sex love'
#tokenized = word_tokenize(query)


# In[232]:

for idx, row in df.iterrows():
    if row['objs']!=0:
        tokenized = word_tokenize(row['objs'])
        pos_tagged = nltk.pos_tag(tokenized)
        satz_synonyms = []
        for wort in pos_tagged:
            if  wort[1] != 'RB' and wort[1] != 'JJ' and wort[1] != 'JJS' and wort[1]!='RBR'and wort[1]!='RBS':
                print(wort[0])
                for syn in wordnet.synsets(wort[0]):
                    for l in syn.lemmas()[:1]:
                        for n in l.name().split():
                            if n not in satz_synonyms:
                                satz_synonyms.append(n)
        else:
            satz_synonyms.append(wort[0])
    #satz_synonyms = [w.split('_') for w in satz_synonyms]
        #print(satz_synonyms[0])
        #print(type(satz_synonyms))
        s = ' '.join(str(v) for v in satz_synonyms)
        s = ''.join(s)
        s = s.split('_')
        s = ' '.join(s)
        print(s)
        df.at[idx,['syn']] = s

df.to_csv('data/q_for_syn.tsv', sep = '\t', index = False)

df
